finding the best aspect ratio + UI designing
i played around with the aspect ratio to see which one fits best. the original is 816:624, but i didnt like it. after scaling, mapping and modelling i found that i liked 1152:648, but its still not resolved becuz i have to find out if it still works for 32x32 tiles instead of 48x48 tiles. there's no tilesets in the game that are 32x32, so i have to design them myself.